Moved Output Helpers to own directory
